Initial commit.
This repository is a stripped down version of a client project that was
intended on being a social network for brokers.

I am removing the social features (as they ran on Parse) and leaving
only the appointment making function of the app.

As this project is almost three years old at the time of me writing
this, I will update old libraries, removed unused code/deprecated code,
and optimize speed.